Add go.mod and go.sum files
This commit adds the go.mod and go.sum files to the repository. The go.mod file specifies the module name and the required versions of the dependencies, while the go.sum file contains the checksums of the dependencies. This ensures that the project can be built with the correct versions of the dependencies.

Signed-off-by: SammyOina <sammyoina@gmail.com>